postgres_to_es/movies_admin/requirements/base.txt
django==3.2
django-model-utils==4.0.0
psycopg2-binary==2.9.1
django-extensions==3.1.3